+++ frontend/src/pages/Methodology/methodologyComponents/KeyTermsAccordion.tsx
@@ -5,18 +5,23 @@ import {
   Paper,
 } from '@mui/material'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+import { parseDescription } from './GlossaryTerm'
 import React, { useState } from 'react'
 import { useIsBreakpointAndUp } from '../../../utils/hooks/useIsBreakpointAndUp'
-import { type DataTypeConfig } from '../../../data/config/MetricConfig'
-import InfoCitations from '../../../reports/ui/InfoCitations'
 import HetTerm from '../../../styles/HetComponents/HetTerm'
 
 interface KeyTermsAccordionProps {
-  datatypeConfigs: DataTypeConfig[]
-  hashId?: string
+  definitionsArray: Array<{
+    topic: string
+    definitions: Array<{
+      key: string
+      description: string
+    }>
+    path?: string
+    id?: string
+  }>
+  id?: string
 }
-
-// TODO: once all new methodology pages are using this component, delete <KeyTerms />  See GH #2821
 
 export default function KeyTermsAccordion(props: KeyTermsAccordionProps) {
   const isMd = useIsBreakpointAndUp('md')
@@ -31,7 +36,7 @@ export default function KeyTermsAccordion(props: KeyTermsAccordionProps) {
   }
 
   return (
-    <div id={props.hashId} className='mt-8 w-full'>
+    <div id={props.id} className='mt-8 w-full'>
       <Paper>
         <Accordion expanded={expanded} onChange={handleAccordionToggle}>
           <AccordionSummary expandIcon={<ExpandMoreIcon />}>
@@ -39,30 +44,23 @@ export default function KeyTermsAccordion(props: KeyTermsAccordionProps) {
           </AccordionSummary>
 
           <AccordionDetails>
-            {props.datatypeConfigs.map((config) => {
+            {props.definitionsArray.map((item) => {
               return (
-                <div className='m-2 mb-8' key={config.dataTypeId}>
-                  <HetTerm>{config.fullDisplayName}</HetTerm>
-                  <p className='mb-1 text-small font-medium'>
-                    Measurement Definition
-                  </p>
-                  <p className='m-0 self-start pt-1 text-small text-altBlack'>
-                    {config.definition?.text}
-                  </p>
-                  <InfoCitations citations={config.definition?.citations} />
-                  {config?.description && (
-                    <>
-                      <p className='mb-1 text-small font-medium'>
-                        Clinical Importance
-                      </p>
-                      <p className='m-0 self-start pt-1 text-small text-altBlack'>
-                        {config.description.text}
-                      </p>
-                      <InfoCitations
-                        citations={config.description?.citations}
-                      />
-                    </>
-                  )}
+                <div className='mb-8' id={item.id} key={item.topic}>
+                  <HetTerm>{item.topic}</HetTerm>
+                  {item.definitions.map((def) => {
+                    return (
+                      <figure
+                        key={def.key}
+                        className='mx-1 mb-2 mt-1 flex flex-col  p-0'
+                      >
+                        <p className='mb-1 text-small font-medium'>{def.key}</p>
+                        <p className='m-0 self-start pt-1 text-small text-altBlack'>
+                          {parseDescription(def.description)}
+                        </p>
+                      </figure>
+                    )
+                  })}
                 </div>
               )
             })}
